+++ CHANGELOG.md
@@ -55,6 +55,7 @@
 - Rename forward to forwardTo
 - Print only suppliers of nodes in doc graphs
 - Plugins can be added and removed now
+- Handle the case that a plugin has plugins
 
 ### Fixed